Original coding was changed to create and cache inverted matrices instead of vector means.
To do this, references to mean were removed, the solve() function replaced the mean(), and the set and get functions were named to indicate they were associated with inverse.

Most of the coding in the original code was not changed.

Comments were added to describe what was happening at each step.